add slp support
This updates the transaction and address views to include
slp properties.  The transaction view adds slp validity, token name,
and token values to any applicable inputs/outputs.  The address view
displays token balances for each token ID.  In the addresses view the
current mint baton location is displayed for a token.  If an NFT child
token is in the address balance then its associated group id is also
displayed.