add slp support (#13)
* add slp support

This updates the transaction and address views to include
slp properties.  The transaction view adds slp validity, token name,
and token values to any applicable inputs/outputs.  The address view
displays token balances for each token ID.  In the addresses view the
current mint baton location is displayed for a token.  If an NFT child
token is in the address balance then its associated group id is also
displayed.

* remove commented block per https://github.com/coinbase/bchd-explorer/pull/13#discussion_r594877748

* show slp address per https://github.com/coinbase/bchd-explorer/pull/13#discussion_r594878561

* fix nit per https://github.com/coinbase/bchd-explorer/pull/13#discussion_r594879280

* rm BCH per https://github.com/coinbase/bchd-explorer/pull/13#discussion_r594879581

* add cash address per https://github.com/coinbase/bchd-explorer/pull/13#discussion_r594878561